debate-engine-py/pyproject.toml
[project]
name = "debate-engine"
version = "0.1.0"
description = "多 Agent 辩论式决策引擎"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langgraph>=0.3.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[project.scripts]
debate = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"

[tool.hatch.build.targets.wheel]
packages = ["src"]
